agent/requirements.txt
# ARETE Agent Core Backend
# FastAPI + LangGraph for AI Technical Interview Platform

# Core Framework
fastapi==0.115.0
uvicorn==0.31.0
websockets==12.0
python-multipart==0.0.9

# LangGraph + LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.19

# OpenAI SDK (for OpenRouter compatibility)
openai==1.54.0

# LiveKit (Video + Voice)
livekit>=0.12.0
livekit-agents==1.3.11
livekit-plugins-deepgram==1.3.11
livekit-plugins-openai==1.3.11
livekit-plugins-silero==1.3.11
livekit-plugins-elevenlabs==1.3.11

# Observability (Arize Phoenix)
arize-phoenix==6.2.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-exporter-otlp==1.28.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10

# Cache
redis==5.2.0

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
httpx==0.27.2
